requirements.txt
alembic==1.13.3
boto3==1.35.34
boto3-stubs==1.35.34
botocore==1.35.34
botocore-stubs==1.35.34
greenlet==3.1.1
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.35
types-awscrt==0.22.0
types-psycopg2==2.9.21.20240819
types-s3transfer==0.10.2
typing_extensions==4.12.2
urllib3==2.2.3
